requirements.txt
acryl-datahub==0.15.0.1
psycopg2-binary==2.9.10
xgboost==3.0.5

# Streaming + logging
confluent-kafka==2.4.0
loguru==0.7.2

# Config
pydantic==2.8.2
pydantic-settings==2.4.0

# ClickHouse client
clickhouse-connect==0.7.17

# Sql alchemy
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
polars==1.33.1
async-timeout==4.0.3
attrs==24.2.0
bentoml==1.4.23
bowler==0.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
dask==2024.8.0
dask-expr==1.1.10
dill==0.3.8
exceptiongroup==1.2.2
fastapi==0.112.2
fastavro==1.9.5
fissix==24.4.24
fsspec==2024.6.1
greenlet==3.0.3
streamlit==1.49.1
grpcio==1.66.0
grpcio-health-checking==1.62.3
mlflow==3.3.2
grpcio-reflection==1.62.3
grpcio-tools==1.62.3
gunicorn==23.0.0
h11==0.14.0
hiredis==2.4.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
importlib-metadata==6.11.0
importlib_resources==6.4.4
Jinja2==3.1.4
matplotlib==3.10.6
seaborn==0.13.2
joblib==1.5.2
scikit-learn==1.7.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
locket==1.0.0
MarkupSafe==2.1.5
mmh3==4.1.0
moreorless==0.4.0
mypy==1.11.2
mypy-extensions==1.0.0
mypy-protobuf==3.1.0
numpy==1.24.4
packaging==24.1
pandas==2.2.2
pandavro==1.5.2
partd==1.4.2
proto-plus==1.24.0
protobuf==4.25.4
psycopg
py4j==0.10.9.7
pyarrow==11.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyspark==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==4.6.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.53
sqlalchemy2-stubs==0.0.2a38
starlette==0.38.2
tabulate==0.9.0
tenacity==8.5.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.5
typeguard==4.3.0
types-protobuf==5.27.0.20240626
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
volatile==2.1.0
watchfiles==0.23.0
websockets==13.0
zipp==3.20.1
feast[redis,kafka]>=0.31.1